Remove dash from action_script for WebUI form submissions
It seems at first glance that 386.12_alpha1 no longer processes
our form submissions for Check for Updates or Refresh Stats due
to the dash in "start_cake-qosupdatecheck" or
"start_cake-qosstatsupdate".

Fix this by removing the dash from the ASP file and adjusting the
service-event-end commands to expect no dash in the passed event
name.